Fix a projection-length bug.
This commit changes the drawing code to only draw projections longer than an
existing first sound. In practice, this means that if the first sound if fully
defined, and the user moves back before the end of the first sound, then the
projection will stick at the end of the first sound rather than getting shorter.