Base determinacy of sound 2 start on sound 1 start.
This changes the state machine for pause1 to match the text of the message: the
beginning of the second sound is indeterminate and ends the simulation if it is
indeterminate with respect to the beginning of the first sound, not the end of
the first sound.

This commit also wraps the blessed_ui commands in an exported function so that
they don't interfere with npm test using mocha.