Port the UI code to keep track of the UI state.
This differs from the previous implementation, which provided individual draw
methods that could be overriden by subclasses of the Draw class. This method
provides better information to subclasses about the nature of the events they
are drawing. For instance, a class can now keep objects that represent each of
the possible UI events and only redraw the ones that have changed.